RED/requirements/base.txt
Django==2.1.3
Pillow==5.3.0
dj-database-url==0.5.0
django-crispy-forms==1.7.2
django-haystack==2.8.1
django-js-asset==1.1.0
django-multiforloop==0.2.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1s
pyelftools==0.25
pylint==1.9.2
python-dateutil==2.7.5
python-decouple==3.1
python-pptx==0.6.5